Split junction exports by file
