+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/getJunctionObjectMetadataIds.ts
@@ -1,8 +1,6 @@
-import {
-  getJunctionConfig,
-  type JunctionObjectMetadataItem,
-  isUsableJunctionConfig,
-} from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
+import { getJunctionConfig } from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
+import { isUsableJunctionConfig } from '@/object-record/record-field/ui/utils/junction/isUsableJunctionConfig';
+import { type JunctionObjectMetadataItem } from '@/object-record/record-field/ui/utils/junction/types/JunctionObjectMetadataItem';
 import { isDefined } from 'twenty-shared/utils';
 
 // A junction object carries no marker of its own, so the relation graph is walked to
