+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/getReverseJunctionConfig.ts
@@ -1,9 +1,7 @@
-import {
-  getJunctionConfig,
-  type JunctionObjectMetadataItem,
-  isUsableJunctionConfig,
-} from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
+import { getJunctionConfig } from '@/object-record/record-field/ui/utils/junction/getJunctionConfig';
 import { getTargetObjectMetadataIdsFromField } from '@/object-record/record-field/ui/utils/junction/getTargetObjectMetadataIdsFromField';
+import { isUsableJunctionConfig } from '@/object-record/record-field/ui/utils/junction/isUsableJunctionConfig';
+import { type JunctionObjectMetadataItem } from '@/object-record/record-field/ui/utils/junction/types/JunctionObjectMetadataItem';
 import { isDefined } from 'twenty-shared/utils';
 
 type ReverseJunctionConfig = {
