+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/isUsableJunctionConfig.ts
@@ -1,0 +1,4 @@
+export const isUsableJunctionConfig = <Config extends { isValid: boolean }>(
+  junctionConfig: Config | null | undefined,
+): junctionConfig is Config & { isValid: true } =>
+  junctionConfig?.isValid === true;
