+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/types/JunctionConfig.ts
@@ -1,0 +1,13 @@
+import { type ValidJunctionConfig } from '@/object-record/record-field/ui/utils/junction/types/ValidJunctionConfig';
+
+type InvalidJunctionConfig = Pick<
+  ValidJunctionConfig,
+  'junctionObjectMetadata'
+> & {
+  targetFields: [];
+  sourceField?: never;
+  isMorphRelation: false;
+  isValid: false;
+};
+
+export type JunctionConfig = ValidJunctionConfig | InvalidJunctionConfig;
