+++ packages/twenty-front/src/modules/object-record/record-field/ui/utils/junction/types/JunctionObjectMetadataItem.ts
@@ -1,0 +1,11 @@
+import { type EnrichedObjectMetadataItem } from '@/object-metadata/types/EnrichedObjectMetadataItem';
+
+export type JunctionObjectMetadataItem = Pick<
+  EnrichedObjectMetadataItem,
+  | 'id'
+  | 'fields'
+  | 'labelIdentifierFieldMetadataId'
+  | 'imageIdentifierFieldMetadataId'
+  | 'nameSingular'
+  | 'namePlural'
+>;
